Move all models to com.keepit.slack.models
