+++ kifi-backend/modules/shoebox/app/com/keepit/slack/models/LibraryToSlackChannel.scala
@@ -1,18 +1,18 @@
-package com.keepit.model
+package com.keepit.slack.models
 
-import com.keepit.common.db.{ States, State, Id }
-import org.joda.time.DateTime
+import com.keepit.common.db.{ Id, State, States }
 import com.keepit.common.time._
+import com.keepit.model.{ Keep, Library }
+import org.joda.time.DateTime
 
 case class LibraryToSlackChannel(
   id: Option[Id[LibraryToSlackChannel]] = None,
   createdAt: DateTime = currentDateTime,
   updatedAt: DateTime = currentDateTime,
   state: State[LibraryToSlackChannel] = LibraryToSlackChannelStates.ACTIVE,
-  ownerId: Id[SlackTeamMembership], // denormalized from SlackIncomingWebhook
-  webhookId: Id[SlackIncomingWebhookInfo],
+  membershipId: Id[SlackTeamMembership],
   library: Id[Library],
-  channel: SlackChannel,
+  webhookId: Id[SlackIncomingWebhookInfo],
   status: SlackIntegrationStatus,
   lastProcessedAt: Option[DateTime],
   lastKeepId: Option[Id[Keep]])
