+++ kifi-backend/modules/shoebox/app/com/keepit/slack/models/SlackChannelToLibrary.scala
@@ -1,15 +1,16 @@
-package com.keepit.model
+package com.keepit.slack.models
 
-import com.keepit.common.db.{ States, State, Id }
-import org.joda.time.DateTime
+import com.keepit.common.db.{ Id, State, States }
 import com.keepit.common.time._
+import com.keepit.model.Library
+import org.joda.time.DateTime
 
 case class SlackChannelToLibrary(
   id: Option[Id[SlackChannelToLibrary]] = None,
   createdAt: DateTime = currentDateTime,
   updatedAt: DateTime = currentDateTime,
   state: State[SlackChannelToLibrary] = SlackChannelToLibraryStates.ACTIVE,
-  ownerId: Id[SlackTeamMembership],
+  membershipId: Id[SlackTeamMembership],
   channel: SlackChannel,
   library: Id[Library],
   status: SlackIntegrationStatus,
